Moves parsing of Map in MapDTO
