+++ lib/webserver/WebServer.js
@@ -13,7 +13,6 @@ const dynamicMiddleware = require("express-dynamic-middleware");
 const basicAuth = require("express-basic-auth");
 const multer = require('multer');
 
-const RRMapParser = require("../RRMapParser");
 const Vacuum = require("../miio/Vacuum");
 
 const LOCAL_UPLOAD_FOLDER = '/mnt/data/valetudo/uploads';
@@ -101,7 +100,7 @@ const WebServer = function (options) {
 
                     if(hashOfNewMap !== self.map.hash) {
                         if (!err) {
-                            const parsedMap = RRMapParser.PARSE(data);
+                            const parsedMap = self.map.PARSE(data);
 
                             self.mapUploadInProgress = false;
